// extension/generated/web/nockchain/public/v2/nockchain.ts
export interface WalletGetBalanceRequest {
  /** wallet address, a base58-encoded cheetah pubkey */
  address?:
    | Base58Pubkey
    | undefined;
  /** tip5 hash identifying the note first-name */
  firstName?: Base58Hash | undefined;
  page: PageRequest | undefined;
}

export interface WalletGetBalanceResponse {
  balance?: Balance | undefined;
  error?: ErrorStatus | undefined;
}

export interface WalletSendTransactionRequest {
  txId: Hash | undefined;
  rawTx: RawTransaction | undefined;
}

export interface WalletSendTransactionResponse {
  ack?: Acknowledged | undefined;
  error?: ErrorStatus | undefined;
}

export interface TransactionAcceptedRequest {
  txId: Base58Hash | undefined;
}

export interface TransactionAcceptedResponse {
  accepted?: boolean | undefined;
  error?: ErrorStatus | undefined;
}

function createBaseWalletGetBalanceRequest(): WalletGetBalanceRequest {
  return { address: undefined, firstName: undefined, page: undefined };
}

export const WalletGetBalanceRequest: MessageFns<WalletGetBalanceRequest> = {
  encode(message: WalletGetBalanceRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.address !== undefined) {
      Base58Pubkey.encode(message.address, writer.uint32(10).fork()).join();
    }
    if (message.firstName !== undefined) {
      Base58Hash.encode(message.firstName, writer.uint32(18).fork()).join();
    }
    if (message.page !== undefined) {
      PageRequest.encode(message.page, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): WalletGetBalanceRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseWalletGetBalanceRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.address = Base58Pubkey.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.firstName = Base58Hash.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.page = PageRequest.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<WalletGetBalanceRequest>): WalletGetBalanceRequest {
    return WalletGetBalanceRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<WalletGetBalanceRequest>): WalletGetBalanceRequest {
    const message = createBaseWalletGetBalanceRequest();
    message.address = (object.address !== undefined && object.address !== null)
      ? Base58Pubkey.fromPartial(object.address)
      : undefined;
    message.firstName = (object.firstName !== undefined && object.firstName !== null)
      ? Base58Hash.fromPartial(object.firstName)
      : undefined;
    message.page = (object.page !== undefined && object.page !== null)
      ? PageRequest.fromPartial(object.page)
      : undefined;
    return message;
  },
};

function createBaseWalletGetBalanceResponse(): WalletGetBalanceResponse {
  return { balance: undefined, error: undefined };
}

export const WalletGetBalanceResponse: MessageFns<WalletGetBalanceResponse> = {
  encode(message: WalletGetBalanceResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.balance !== undefined) {
      Balance.encode(message.balance, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      ErrorStatus.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): WalletGetBalanceResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseWalletGetBalanceResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.balance = Balance.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = ErrorStatus.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<WalletGetBalanceResponse>): WalletGetBalanceResponse {
    return WalletGetBalanceResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<WalletGetBalanceResponse>): WalletGetBalanceResponse {
    const message = createBaseWalletGetBalanceResponse();
    message.balance = (object.balance !== undefined && object.balance !== null)
      ? Balance.fromPartial(object.balance)
      : undefined;
    message.error = (object.error !== undefined && object.error !== null)
      ? ErrorStatus.fromPartial(object.error)
      : undefined;
    return message;
  },
};

function createBaseWalletSendTransactionRequest(): WalletSendTransactionRequest {
  return { txId: undefined, rawTx: undefined };
}

export const WalletSendTransactionRequest: MessageFns<WalletSendTransactionRequest> = {
  encode(message: WalletSendTransactionRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.txId !== undefined) {
      Hash.encode(message.txId, writer.uint32(10).fork()).join();
    }
    if (message.rawTx !== undefined) {
      RawTransaction.encode(message.rawTx, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): WalletSendTransactionRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseWalletSendTransactionRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.txId = Hash.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.rawTx = RawTransaction.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<WalletSendTransactionRequest>): WalletSendTransactionRequest {
    return WalletSendTransactionRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<WalletSendTransactionRequest>): WalletSendTransactionRequest {
    const message = createBaseWalletSendTransactionRequest();
    message.txId = (object.txId !== undefined && object.txId !== null) ? Hash.fromPartial(object.txId) : undefined;
    message.rawTx = (object.rawTx !== undefined && object.rawTx !== null)
      ? RawTransaction.fromPartial(object.rawTx)
      : undefined;
    return message;
  },
};

function createBaseWalletSendTransactionResponse(): WalletSendTransactionResponse {
  return { ack: undefined, error: undefined };
}

export const WalletSendTransactionResponse: MessageFns<WalletSendTransactionResponse> = {
  encode(message: WalletSendTransactionResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.ack !== undefined) {
      Acknowledged.encode(message.ack, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      ErrorStatus.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): WalletSendTransactionResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseWalletSendTransactionResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.ack = Acknowledged.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = ErrorStatus.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<WalletSendTransactionResponse>): WalletSendTransactionResponse {
    return WalletSendTransactionResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<WalletSendTransactionResponse>): WalletSendTransactionResponse {
    const message = createBaseWalletSendTransactionResponse();
    message.ack = (object.ack !== undefined && object.ack !== null) ? Acknowledged.fromPartial(object.ack) : undefined;
    message.error = (object.error !== undefined && object.error !== null)
      ? ErrorStatus.fromPartial(object.error)
      : undefined;
    return message;
  },
};

function createBaseTransactionAcceptedRequest(): TransactionAcceptedRequest {
  return { txId: undefined };
}

export const TransactionAcceptedRequest: MessageFns<TransactionAcceptedRequest> = {
  encode(message: TransactionAcceptedRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.txId !== undefined) {
      Base58Hash.encode(message.txId, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TransactionAcceptedRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTransactionAcceptedRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.txId = Base58Hash.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<TransactionAcceptedRequest>): TransactionAcceptedRequest {
    return TransactionAcceptedRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TransactionAcceptedRequest>): TransactionAcceptedRequest {
    const message = createBaseTransactionAcceptedRequest();
    message.txId = (object.txId !== undefined && object.txId !== null)
      ? Base58Hash.fromPartial(object.txId)
      : undefined;
    return message;
  },
};
// ...
export type NockchainServiceDefinition = typeof NockchainServiceDefinition;
export const NockchainServiceDefinition = {
  name: "NockchainService",
  fullName: "nockchain.public.v2.NockchainService",
  methods: {
    walletGetBalance: {
      name: "WalletGetBalance",
      requestType: WalletGetBalanceRequest,
      requestStream: false,
      responseType: WalletGetBalanceResponse,
      responseStream: false,
      options: {},
    },
    walletSendTransaction: {
      name: "WalletSendTransaction",
      requestType: WalletSendTransactionRequest,
      requestStream: false,
      responseType: WalletSendTransactionResponse,
      responseStream: false,
      options: {},
    },
    /**
     * rpc TransactionConfirmation(TransactionConfirmationRequest)
     *    returns (TransactionConfirmationResponse);
     */
    transactionAccepted: {
      name: "TransactionAccepted",
      requestType: TransactionAcceptedRequest,
      requestStream: false,
      responseType: TransactionAcceptedResponse,
      responseStream: false,
      options: {},
    },
  },
} as const;

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}